Display Complete Metadata Associated with Soil Thickness Data in the R Console

Displays the complete set of metadata associated with the soil thickness data in your R console. For including the metadata in documents or other methods outside of R, see get_soil_thickness for an example using pander::pander to print the metadata.

Usage

display_soil_thickness_metadata(x)

Arguments

x

An abares.soil.thickness.files object

Value

Nothing, called for its side effects, it prints the complete metadata file to the R console

See also

Examples

get_soil_thickness(cache = TRUE) |>
display_soil_thickness_metadata()
#> 
#> ── Soil Thickness for Australian areas of intensive agriculture of Layer 1 (A Ho
#> 
#> ── Dataset ANZLIC ID ANZCW1202000149 ──
#> 
#> Custodian CSIRO, Land & Water
#> 
#> Jurisdiction Australia
#> 
#> Description Abstract Surface of predicted Thickness of soil layer 1 (A Horizon
#> - top-soil) surface for the intensive agricultural areas of Australia.  Data
#> modelled from area based observations made by soil agencies both State and
#> CSIRO and presented as .0.01 degree grid cells.
#> 
#> Topsoils (A horizons) are defined as the surface soil layers in which organic
#> matter accumulates, and may include dominantly organic surface layers (O and P
#> horizons).
#> 
#> The depth of topsoil is important because, with their higher organic matter
#> contents, topsoils (A horizon) generally have more suitable properties for
#> agriculture, including higher permeability and higher levels of soil nutrients.
#> 
#> Estimates of soil depths are needed to calculate the amount of any soil
#> constituent in either volume or mass terms (bulk density is also needed) - for
#> example, the volume of water stored in the rooting zone potentially available
#> for plant use, to assess total stores of soil carbon for Greenhouse inventory
#> or to assess total stores of nutrients.
#> 
#> The pattern of soil depth is strongly related to topography - the shape and
#> slope of the land.  Deeper soils are typically found in the river valleys where
#> soils accumulate on floodplains and at the footslopes of ranges (zones of
#> deposition), while soils on hillslopes (zones of erosion) tend to be shallow.
#> Map of thickness of topsoil was derived from soil map data and interpreted
#> tables of soil properties for specific soil groups.
#> 
#> The quality of data on soil depth in existing soil profile datasets is
#> questionable and as the thickness of soil horizons varies locally with
#> topography, values for map units are general averages.
#> 
#> The final ASRIS polygon attributed surfaces are a mosaic of all of the data
#> obtained from various state and federal agencies. The surfaces have been
#> constructed with the best available soil survey information available at the
#> time. The surfaces also rely on a number of assumptions. One being that an area
#> weighted mean is a good estimate of the soil attributes for that polygon or
#> map-unit. Another assumption made is that the look-up tables provided by
#> McKenzie et al. (2000), state and territories accurately depict the soil
#> attribute values for each soil type.
#> 
#> The accuracy of the maps is most dependent on the scale of the original polygon
#> data sets and the level of soil survey that has taken place in each state.  The
#> scale of the various soil maps used in deriving this map is available by
#> accessing the data-source grid, the scale is used as an assessment of the
#> likely accuracy of the modelling.  The Atlas of Australian Soils is considered
#> to be the least accurate dataset and has therefore only been used where there
#> is no state based data.  Of the state datasets Western Australian sub-systems,
#> South Australian land systems and NSW soil landscapes and reconnaissance
#> mapping would be the most reliable based on scale. NSW soil landscapes and
#> reconnaissance mapping use only one dominant soil type per polygon in the
#> estimation of attributes.  South Australia and Western Australia use several
#> soil types per polygon or map-unit.
#> 
#> The digital map data is provided in geographical coordinates based on the World
#> Geodetic System 1984 (WGS84) datum. This raster data set has a grid resolution
#> of 0.001 degrees (approximately equivalent to 1.1 km).
#> 
#> The data set is a product of the National Land and Water Resources Audit
#> (NLWRA) as a base dataset.
#> 
#> Search Word(s) AGRICULTURE SOIL Physics Models
#> 
#> Geographic Extent Name(s) GEN Category
#> 
#> GEN Custodial Jurisdiction
#> 
#> GEN Name
#> 
#> Geographic Bounding Box North Bounding Latitude -10.707149 South Bounding
#> Latitude -43.516831 East Bounding Longitude 113.19673 West Bounding Longitude
#> 153.990779
#> 
#> Geographic Extent Polygon(s) 115.0 -33.5,115.7 -33.3,115.7 -31.7,113.2
#> -26.2,113.5 -25.4,114.1 -26.4,114.3 -26.0,113.4 -24.3,114.1 -21.8,122.3
#> -18.2,122.2 -17.2,126.7 -13.6,129.1 -14.9,130.6 -12.3,132.6 -12.1,132.5
#> -11.6,131.9 -11.3,132.0 -11.1,137.0 -12.2,135.4 -14.7,140.0 -17.7,140.8
#> -17.4,141.7 -15.1,141.4 -13.7,142.2 -10.9,142.7 -10.7,143.9 -14.5,144.6
#> -14.1,145.3 -14.9,146.3 -18.8,148.9 -20.5,150.9 -22.6,153.2 -25.9,153.7
#> -28.8,153.0 -31.3,150.8 -34.8,150.0 -37.5,147.8 -37.9,146.3 -39.0,144.7
#> -38.4,143.5 -38.8,141.3 -38.4,139.7 -37.3,139.7 -36.9,139.9 -36.7,138.9
#> -35.5,138.1 -35.7,138.6 -34.7,138.1 -34.2,137.8 -35.1,136.9 -35.3,137.0
#> -34.9,137.5 -34.9,137.4 -34.0,137.9 -33.5,137.8 -32.6,137.3 -33.6,135.9
#> -34.7,136.1 -34.8,136.0 -35.0,135.1 -34.6,135.2 -34.5,135.4 -34.5,134.7
#> -33.3,134.0 -32.9,133.7 -32.1,133.3 -32.2,132.2 -32.0,131.3 -31.5,127.3
#> -32.3,126.0 -32.3,123.6 -33.9,123.2 -34.0,122.1 -34.0,121.9 -33.8,119.9
#> -34.0,119.6 -34.4,118.0 -35.1,116.0 -34.8,115.0 -34.3,115.0 -33.5
#> 
#> 147.8 -42.9,147.9 -42.6,148.2 -42.1,148.3 -42.3,148.3 -41.3,148.3 -41.0,148.0
#> -40.7,147.4 -41.0,146.7 -41.1,146.6 -41.2,146.5 -41.1,146.4 -41.2,145.3
#> -40.8,145.3 -40.7,145.2 -40.8,145.2 -40.8,145.2 -40.8,145.0 -40.8,144.7
#> -40.7,144.7 -41.2,145.2 -42.2,145.4 -42.2,145.5 -42.4,145.5 -42.5,145.2
#> -42.3,145.5 -43.0,146.0 -43.3,146.0 -43.6,146.9 -43.6,146.9 -43.5,147.1
#> -43.3,147.0 -43.1,147.2 -43.3,147.3 -42.8,147.4 -42.9,147.6 -42.8,147.5
#> -42.8,147.8 -42.9,147.9 -43.0,147.7 -43.0,147.8 -43.2,147.9 -43.2,147.9
#> -43.2,148.0 -43.2,148.0 -43.1,148.0 -42.9,147.8 -42.9
#> 
#> 136.7 -13.8,136.7 -13.7,136.6 -13.7,136.6 -13.8,136.4 -13.8,136.4 -14.1,136.3
#> -14.2,136.9 -14.3,137.0 -14.2,136.9 -14.2,136.7 -14.1,136.9 -13.8,136.7
#> -13.8,136.7 -13.8
#> 
#> 139.5 -16.6,139.7 -16.5,139.4 -16.5,139.2 -16.7,139.3 -16.7,139.5 -16.6
#> 
#> 153.0 -25.2,153.0 -25.7,153.1 -25.8,153.4 -25.0,153.2 -24.7,153.2 -25.0,153.0
#> -25.2
#> 
#> 137.5 -36.1,137.7 -35.9,138.1 -35.9,137.9 -35.7,137.6 -35.7,137.6 -35.6,136.6
#> -35.8,136.7 -36.1,137.2 -36.0,137.5 -36.1
#> 
#> 143.9 -39.7,144.0 -39.6,144.1 -39.8,143.9 -40.2,143.9 -40.0,143.9 -39.7
#> 
#> 148.0 -39.7,147.7 -39.9,147.9 -39.9,148.0 -40.1,148.1 -40.3,148.3 -40.2,148.3
#> -40.0,148.0 -39.7
#> 
#> 148.1 -40.4,148.0 -40.4,148.4 -40.3,148.4 -40.5,148.1 -40.4
#> 
#> 130.4 -11.3,130.4 -11.2,130.6 -11.3,130.7 -11.4,130.9 -11.3,131.0 -11.4,131.1
#> -11.3,131.2 -11.4,131.3 -11.2,131.5 -11.4,131.5 -11.5,131.0 -11.9,130.8
#> -11.8,130.6 -11.7,130.0 -11.8,130.1 -11.7,130.3 -11.7,130.1 -11.5,130.4 -11.3
#> 
#> Data Currency Beginning date 1999-09-01
#> 
#> Ending date 2001-03-31
#> 
#> Dataset Status Progress COMPLETE
#> 
#> Maintenance and Update Frequency NOT PLANNED
#> 
#> Access Stored Data Format DIGITAL - ESRI Arc/Info integer GRID
#> 
#> Available Format Type DIGITAL - ESRI Arc/Info integer GRID
#> 
#> Access Constraint Subject to the terms & condition of the data access &
#> management agreement between the National Land & Water Audit and ANZLIC parties
#> 
#> Data Quality Lineage The soil attribute surface was created using the following
#> datasets 1. The digital polygon coverage of the Soil-Landforms of the Murray
#> Darling Basis (MDBSIS)(Bui et al. 1998), classified as principal profile forms
#> (PPF's) (Northcote 1979). 2. The digital Atlas of Australian Soils (Northcote
#> et al.1960-1968)(Leahy, 1993). 3. Western Australia land systems coverage
#> (Agriculture WA). 4. Western Australia sub-systems coverage (Agriculture WA).
#> 5. Ord river catchment soils coverage (Agriculture WA). 6. Victoria soils
#> coverage (Victorian Department of Natural Resources and Environment - NRE). 7.
#> NSW Soil Landscapes and reconnaissance soil landscape mapping (NSW Department
#> of Land and Water Conservation - DLWC). 8. New South Wales Land systems west
#> (NSW Department of Land and Water Conservation - DLWC). 9. South Australia soil
#> land-systems (Primary Industries and Resources South Australia - PIRSA). 10.
#> Northern Territory soils coverage (Northern Territory Department of Lands,
#> Planning and Environment). 11. A mosaic of Queensland soils coverages
#> (Queensland Department of Natural Resources - QDNR). 12. A look-up table
#> linking PPF values from the Atlas of Australian Soils with interpreted soil
#> attributes (McKenzie et al. 2000). 13. Look_up tables provided by WA
#> Agriculture linking WA soil groups with interpreted soil attributes. 14.
#> Look_up tables provided by PIRSA linking SA soil groups with interpreted soil
#> attributes.
#> 
#> The continuous raster surface representing Thickness of soil layer 1 was
#> created by combining national and state level digitised land systems maps and
#> soil surveys linked to look-up tables listing soil type and corresponding
#> attribute values.
#> 
#> Because thickness is used sparingly in the Factual Key, estimations of
#> thickness in the look-up tables were made using empirical correlations for
#> particular soil types.
#> 
#> To estimate a soil attribute where more than one soil type was given for a
#> polygon or map-unit, the soil attribute values related to each soil type in the
#> look-up table were weighted according to the area occupied by that soil type
#> within the polygon or map-unit. The final soil attribute values are an area
#> weighted average for a polygon or map-unit.  The polygon data was then
#> converted to a continuous raster surface using the soil attribute values
#> calculated for each polygon.
#> 
#> The ASRIS soil attribute surfaces created using polygon attribution relied on a
#> number of data sets from various state agencies.  Each polygon data set was
#> turned into a continuous surface grid based on the calculated soil attribute
#> value for that polygon.  The grids where then merged on the basis that, where
#> available, state data replaced the Atlas of Australian Soils and MDBSIS.
#> MDBSIS derived soil attribute values were restricted to areas where MDBSIS was
#> deemed to be more accurate that the Atlas of Australian Soils (see Carlile et
#> al (2001a).
#> 
#> In cases where a soil type was missing from the look-up table or layer 2 did
#> not exist for that soil type, the percent area of the soils remaining were
#> adjusted prior to calculating the final soil attribute value. The method used
#> to attribute polygons was dependent on the data supplied by individual State
#> agencies.
#> 
#> The modelled grid was resampled from 0.0025 degree cells to 0.01 degree cells
#> using bilinear interpolation
#> 
#> Positional Accuracy The predictive surface is a 0.01 X 0.01 degree grid and has
#> a locational accurate of about 1m.
#> 
#> The positional accuracy of the defining polygons have variable positional
#> accuracy most locations are expected to be within 100m of the recorded
#> location.  The vertical accuracy is not relevant. The positional assessment has
#> been made by considering the tools used to generate the locational information
#> and contacting the data providers.
#> 
#> The other parameters used in the production of the led surface have a range of
#> positional accuracy ranging from + - 50 m to + - kilometres.  This contribute
#> to the loss of attribute accuracy in the surface.
#> 
#> Attribute Accuracy Input attribute accuracy for the areas is highly variable.
#> The predictive has a variable and much lower attribute accuracy due to the
#> irregular distribution and the limited positional accuracy of the parameters
#> used for modelling.
#> 
#> There are several sources of error in estimating soil depth and thickness of
#> horizons for the look-up tables.  Because thickness is used sparingly in the
#> Factual Key, estimations of thickness in the look-up tables were made using
#> empirical correlations for particular soil types.  The quality of data on soil
#> depth in existing soil profile datasets is questionable, in soil mapping,
#> thickness of soil horizons varies locally with topography, so values for map
#> units are general averages.  The definition of the depth of soil or regolith is
#> imprecise and it can be difficult to determine the lower limit of soil.
#> 
#> The assumption made that an area weighted mean of soil attribute values based
#> on soil type is a good estimation of a soil property is debatable, in that it
#> does not supply the soil attribute value at any given location. Rather it is
#> designed to show national and regional patterns in soil properties. The use of
#> the surfaces at farm or catchment scale modelling may prove inaccurate. Also
#> the use of look-up tables to attribute soil types is only as accurate as the
#> number of observations used to estimate a attribute value for a soil type. Some
#> soil types in the look-up tables may have few observations, yet the average
#> attribute value is still taken as the attribute value for that soil type.
#> Different states are using different taxonomic schemes making a national soil
#> database difficult. Another downfall of the area weighted approach is that some
#> soil types may not be listed in look-up tables. If a soil type is a dominant
#> one within a polygon or map-unit, but is not listed within the look-up table or
#> is not attributed within the look-up table then the final soil attribute value
#> for that polygon will be biased towards the minor soil types that do exist.
#> This may also happen when a large area is occupied by a soil type which has no
#> B horizon. In this case the final soil attribute value will be area weighted on
#> the soils with a B horizon, ignoring a major soil type within that polygon or
#> map-unit. The layer 2 surfaces have large areas of no-data because all soils
#> listed for a particular map-unit or polygon had no B horizon.
#> 
#> Logical Consistency Surface is fully logically consistent as only one parameter
#> is shown, as predicted average Soil Thickness within each grid cell
#> 
#> Completeness Surface is nearly complete.  There are some areas (about %1
#> missing) for which insufficient parameters were known to provide a useful
#> prediction and thus attributes are absent in these areas.
#> 
#> Contact Information Contact Organisation (s) CSIRO, Land & Water
#> 
#> Contact Position Project Leader
#> 
#> Mail Address ACLEP, GPO 1666
#> 
#> Locality Canberra
#> 
#> State ACT
#> 
#> Country AUSTRALIA
#> 
#> Postcode 2601
#> 
#> Telephone 02 6246 5922
#> 
#> Facsimile 02 6246 5965
#> 
#> Electronic Mail Address neil.mckenzie@cbr.clw.csiro.au
#> 
#> Metadata Date Metadata Date 2001-07-01
#> 
#> Additional Metadata Additional Metadata
#> 
#> Entity and Attributes Entity Name Soil Thickness Layer 1 (derived from mapping)
#> 
#> Entity description Estimated Soil Thickness (mm) of Layer 1 on a cell by cell
#> basis
#> 
#> Feature attribute name VALUE
#> 
#> Feature attribute definition Predicted average Thickness (mm) of soil layer 1
#> in the 0.01 X 0.01 degree quadrat
#> 
#> Data Type Spatial representation type RASTER
#> 
#> Projection Map projection GEOGRAPHIC
#> 
#> Datum WGS84
#> 
#> Map units DECIMAL DEGREES
#> 
#> Scale Scale/ resolution 1:1 000 000
#> 
#> Usage Purpose Estimates of soil depths are needed to calculate the amount of
#> any soil constituent in either volume or mass terms (bulk density is also
#> needed) - for example, the volume of water stored in the rooting zone
#> potentially available for plant use, to assess total stores of soil carbon for
#> Greenhouse inventory or to assess total stores of nutrients.
#> 
#> Provide indications of probable Thickness soil layer 1 in agricultural areas
#> where soil thickness testing has not been carried out
#> 
#> Use Use Limitation This dataset is bound by the requirements set down by the
#> National Land & Water Resources Audit
#>