Fetch Australian Gridded Farm Data

Fetch Australian Gridded Farm Data

Usage

get_agfd(fixed_prices = TRUE, cache = TRUE)

Arguments

fixed_prices

Boolean Download historical climate and prices or historical climate and fixed prices as described in (Hughes et al. 2022). Defaults to TRUE and downloads the data with historical climate and fixed prices “to isolate the effects of climate variability on financial incomes for broadacre farm businesses” (ABARES 2024). Using TRUE will download simulations where global output and input price indexes are fixed at values from the most recently completed financial year.

cache

Boolean Cache the Australian Gridded Farm Data files after download using tools::R_user_dir to identify the proper directory for storing user data in a cache for this package. Defaults to TRUE, caching the files locally. If FALSE, this function uses tempdir() and the files are deleted upon closing of the R session.

Value

A list of NetCDF files containing the Australian Gridded Farm Data with fullnames, i.e., with the file path prepended to the file name

References

Australian gridded farm data, Australian Bureau of Agricultural and Resource Economics and Sciences, Canberra, July 2024, DOI: 10.25814/7n6z-ev41. CC BY 4.0.

N. Hughes, W.Y. Soh, C. Boult, K. Lawson, Defining drought from the perspective of Australian farmers, Climate Risk Management, Volume 35, 2022, 100420, ISSN 2212-0963, DOI: 10.1016/j.crm.2022.100420.

Examples

get_agfd()
#> Warning: restarting interrupted promise evaluation
#> Warning: internal error -3 in R_decompress1
#> Error in eval(expr, envir, enclos): lazy-load database '/Users/adamsparks/Library/R/arm64/4.4/library/abares/R/abares.rdb' is corrupt