Download Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1

Download Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1

Usage

get_soil_thickness(cache = TRUE)

Arguments

cache

Boolean Cache the soil thickness data files after download using tools::R_user_dir to identify the proper directory for storing user data in a cache for this package. Defaults to TRUE, caching the files locally. If FALSE, this function uses tempdir() and the files are deleted upon closing of the R session.

A custom print method is provided that will print the metadata associated with these data. Examples are provided for interacting with the metadata directly.

Value

An abares.soil.thickness object, which is a named list with the file path of the resulting ESRI Grid file and text file of metadata

References

https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6

See also

Examples

x <- get_soil_thickness()
#> Error in list_files(file, ignore_missing, TRUE, verbosity): File '' does not exist.

# View the metadata with pretty printing
x
#> Error: object 'x' not found

# Extract the metadata as an object in your R session and use it with
# {pander}, useful for Markdown files

library(pander)
y <- x$metadata
#> Error: object 'x' not found
pander(y)
#> Error: object 'y' not found